requirements.txt
requests
beautifulsoup4
selenium
webdriver-manager
pyinputplus
pandas
streamlit
schedule
python-dotenv
